video: tegra: host: Merge job allocs
In nvhost_job, we allocate the job structure via vzalloc, and pin_ids
via kmalloc. Merge the latter to the former so that we do only one
allocation.

As memory ids received from user space are u32, change the types of
functions to match. Also call the validating version of
nvmap_duplicate_handle_id().

Bug 1231756

Reviewed-on: http://git-master/r/204463
Tested-by: Harshada Kale <hkale@nvidia.com>

+++ drivers/video/tegra/host/chip_support.h
@@ -156,7 +156,7 @@ struct nvhost_mem_ops {
 	void (*kunmap)(struct mem_handle *, unsigned int, void *);
 	int (*pin_array_ids)(struct mem_mgr *,
 			struct platform_device *,
-			long unsigned *,
+			u32 *,
 			dma_addr_t *,
 			u32,
 			struct nvhost_job_unpin *);
