+++ drivers/video/tegra/host/dmabuf.h
@@ -42,9 +42,9 @@ void nvhost_dmabuf_kunmap(struct mem_handle *handle, unsigned int pagenum,
 		void *addr);
 int nvhost_dmabuf_get(u32 id, struct platform_device *dev);
 int nvhost_dmabuf_pin_array_ids(struct platform_device *dev,
-		long unsigned *ids,
-		long unsigned id_type_mask,
-		long unsigned id_type,
+		u32 *ids,
+		u32 id_type_mask,
+		u32 id_type,
 		u32 count,
 		struct nvhost_job_unpin *unpin_data,
 		dma_addr_t *phys_addr);
