+++ drivers/video/tegra/host/nvhost_memmgr.h
@@ -70,7 +70,7 @@ u32 nvhost_memmgr_handle_to_id(struct mem_handle *handle);
 
 int nvhost_memmgr_pin_array_ids(struct mem_mgr *mgr,
 		struct platform_device *dev,
-		long unsigned *ids,
+		u32 *ids,
 		dma_addr_t *phys_addr,
 		u32 count,
 		struct nvhost_job_unpin *unpin_data);
